feat: operations support + refactoring
re #10
