+++ test/samples/monOps.json
@@ -1,0 +1,6 @@
+{
+  "name": "sample #26",
+  "type": "JSON",
+  "error": "unary operations",
+  "version": ~6
+}
